requirements.txt
torch
torchvision
numpy
scikit-learn
pillow
opencv-python
tqdm
streamlit
